telegram/requirements.txt
# Telegram Bot Integration for CX Agent Studio
# Python dependencies

# Telegram Bot API
python-telegram-bot>=20.7

# Google Cloud Authentication
google-auth>=2.25.0

# HTTP requests for REST API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0
